# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Car auction app",
        "title": "AuthServiceAPI",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
        "/car": {
            "get": {
                "description": "Get car list from auction",
                "produces": [
                    "application/json"
                ],
                "summary": "Get car list",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/httpmodels.CarResponse"
                            }
                        }
                    },
                    "500": {
                        "description": "[ITERNAL_SERVER_ERROR]",
                        "schema": {
                            "$ref": "#/definitions/httpmodels.CustomHttpError"
                        }
                    }
                }
            },
            "post": {
                "description": "Create car for auction",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Create car",
                "parameters": [
                    {
                        "description": "request body",
                        "name": "params",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/httpmodels.CarCreateRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created"
                    },
                    "400": {
                        "description": "[INCORRECT_REQUEST_BODY], [VALIDATION_FAILED]",
                        "schema": {
                            "$ref": "#/definitions/httpmodels.CustomHttpError"
                        }
                    },
                    "500": {
                        "description": "[ITERNAL_SERVER_ERROR]",
                        "schema": {
                            "$ref": "#/definitions/httpmodels.CustomHttpError"
                        }
                    }
                }
            }
        },
        "/car/{carId}": {
            "get": {
                "description": "Get car from auction",
                "produces": [
                    "application/json"
                ],
                "summary": "Get car",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Id of the car",
                        "name": "carId",
                        "in": "path"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/httpmodels.CarResponse"
                        }
                    },
                    "404": {
                        "description": "[NOT_FOUND]",
                        "schema": {
                            "$ref": "#/definitions/httpmodels.CustomHttpError"
                        }
                    },
                    "500": {
                        "description": "[ITERNAL_SERVER_ERROR]",
                        "schema": {
                            "$ref": "#/definitions/httpmodels.CustomHttpError"
                        }
                    }
                }
            },
            "put": {
                "description": "Update car for auction",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Update car",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Id of the car",
                        "name": "carId",
                        "in": "path"
                    },
                    {
                        "description": "request body",
                        "name": "params",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/httpmodels.CarCreateRequest"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "[INCORRECT_REQUEST_BODY], [VALIDATION_FAILED]",
                        "schema": {
                            "$ref": "#/definitions/httpmodels.CustomHttpError"
                        }
                    },
                    "500": {
                        "description": "[ITERNAL_SERVER_ERROR]",
                        "schema": {
                            "$ref": "#/definitions/httpmodels.CustomHttpError"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete car from auction",
                "summary": "Delete car",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Id of the car",
                        "name": "carId",
                        "in": "path"
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "500": {
                        "description": "[ITERNAL_SERVER_ERROR]",
                        "schema": {
                            "$ref": "#/definitions/httpmodels.CustomHttpError"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "httpmodels.CarCreateRequest": {
            "type": "object",
            "required": [
                "color",
                "max_speed_kmp",
                "model_name",
                "price_in_cents",
                "vendor_name"
            ],
            "properties": {
                "color": {
                    "type": "string",
                    "maxLength": 255,
                    "example": "#ffb7d5"
                },
                "max_speed_kmp": {
                    "type": "integer",
                    "minimum": 0,
                    "example": 224
                },
                "max_speed_mph": {
                    "type": "integer",
                    "minimum": 0,
                    "example": 140
                },
                "model_name": {
                    "type": "string",
                    "maxLength": 255,
                    "minLength": 2,
                    "example": "Mx-5"
                },
                "price_in_cents": {
                    "type": "integer",
                    "minimum": 0,
                    "example": 1000000
                },
                "vendor_name": {
                    "type": "string",
                    "maxLength": 255,
                    "minLength": 2,
                    "example": "Mazda"
                }
            }
        },
        "httpmodels.CarResponse": {
            "type": "object",
            "required": [
                "color",
                "max_speed_kmp",
                "model_name",
                "price_in_cents",
                "vendor_name"
            ],
            "properties": {
                "color": {
                    "type": "string",
                    "maxLength": 255,
                    "example": "#ffb7d5"
                },
                "date_created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "max_speed_kmp": {
                    "type": "integer",
                    "minimum": 0,
                    "example": 224
                },
                "max_speed_mph": {
                    "type": "integer",
                    "minimum": 0,
                    "example": 140
                },
                "model_name": {
                    "type": "string",
                    "maxLength": 255,
                    "minLength": 2,
                    "example": "Mx-5"
                },
                "price_in_cents": {
                    "type": "integer",
                    "minimum": 0,
                    "example": 1000000
                },
                "vendor_name": {
                    "type": "string",
                    "maxLength": 255,
                    "minLength": 2,
                    "example": "Mazda"
                }
            }
        },
        "httpmodels.CustomHttpError": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        }
    }
}